Fixed delete request
