+++ app/javascript/react/components/reviews/ReviewList.js
@@ -2,32 +2,18 @@ import React from "react";
 import ReviewTile from "./ReviewTile";
 
 const ReviewList = (props) => {
-  const { reviews, animal } = props;
-
-  const deleteReview = async (reviewId) => {
-    try {
-      const response = await fetch(`/api/v1/animals/${animal}/reviews/${reviewId}`, {
-        credentials: "same-origin",
-        method: "DELETE",
-        headers: {
-          Accept: "application/json",
-          "Content-Type": "application/json",
-        }
-      })
-    } catch(err) {
-      console.error(err)
-    }
-  }
+  const { reviews, animal, deleteReview } = props;
 
   let reviewList = reviews.map((review) => {
+    debugger
     return (
       <ReviewTile
         key={review.id}
         title={review.title}
         description={review.description}
         rating={review.rating}
-        deleteReview={() => deleteReview(review.id)}
-        user="placeholder person"
+        deleteReview={() => deleteReview(review.id, animal.id)}
+        user={review.user_id}
       />
     );
   });
